+++ test/mpfr.jl
@@ -895,7 +895,7 @@ end
         ends::String="",
         starts::String="")
         sx = sprint(show, x)
-        scx = sprint(showcompact, x)
+        scx = sprint(show, x, context=:compact => true)
         strx = string(x)
         @test sx == strx
         @test length(scx) < 20
@@ -915,8 +915,8 @@ end
     test_show_bigfloat(big"-2.3457645687563543266576889678956787e-10000", starts="-2.345", ends="e-10000")
 
     for to_string in [string,
-        x->sprint(show, x),
-        x->sprint(showcompact,x)]
+                      x->sprint(show, x),
+                      x->sprint(show, x, context=:compact => true)]
         @test to_string(big"0.0") == "0.0"
         @test to_string(big"-0.0") == "-0.0"
         @test to_string(big"1.0") == "1.0"