+++ NEWS.md
@@ -1093,7 +1093,11 @@ Deprecated or removed
   * `DevNull`, `STDIN`, `STDOUT`, and `STDERR` have been renamed to `devnull`, `stdin`, `stdout`,
     and `stderr`, respectively ([#25786]).
 
-  * `wait` and `fetch` on `Task` now resemble the interface of `Future`
+  * `wait` and `fetch` on `Task` now resemble the interface of `Future`.
+
+  * `showcompact(io, x...)` has been deprecated in favor of
+    `show(IOContext(io, :compact => true), x...)` ([#26080]).
+    Use `sprint(show, x..., context=:compact => true)` instead of `sprint(showcompact, x...)`.
 
 Command-line option changes
 ---------------------------
@@ -1391,4 +1395,5 @@ Command-line option changes
 [#25998]: https://github.com/JuliaLang/julia/issues/25998
 [#26009]: https://github.com/JuliaLang/julia/issues/26009
 [#26071]: https://github.com/JuliaLang/julia/issues/26071
+[#26080]: https://github.com/JuliaLang/julia/issues/26080
 [#26149]: https://github.com/JuliaLang/julia/issues/26149